+++ CHANGELOG.md
@@ -13,6 +13,8 @@ FEATURES
 - Add `DefaultTemplateSid` input field to service create and update input structs
 - Add `Totp` response struct to service create, fetch and update response structs
 - Add `Totp` input struct to service create and update input structs
+- Add `TemplateSid` input field to verification create input structs
+- Add `TemplateCustomSubstitutions` input field to verification create input structs
 
 ## v0.19.0 (2021-09-05)